Module Gr.Print

val beginp : string -> unit

beginp path opens and activates a print device.

This function opens an additional graphics output device. The device type is obtained from the given file extension.

Parameters

pathname: Filename for the print device.

The following file types are supported:

.ps, .eps PostScript

.pdf Portable Document Format

.bmp Windows Bitmap (BMP)

.jpeg, .jpg JPEG image file

.png Portable Network Graphics file (PNG)

.tiff, .tif Tagged Image File Format (TIFF)

.svg Scalable Vector Graphics

.wmf Windows Metafile

.mp4 MPEG-4 video file

.webm WebM video file

.ogg Ogg video file

val endp : unit -> unit

val beginp_extended : string -> [< `Color | `GreyScale ] -> [< `A0 | `A1 | `A2 | `A3 | `A4 | `A5 | `A6 | `A7 | `A8 | `A9 | `B0 | `B1 | `B10 | `B2 | `B3 | `B4 | `B5 | `B6 | `B7 | `B8 | `B9 | `C5E | `Comm10E | `DLE | `Executive | `Folio | `Ledger | `Legal | `Letter | `Tabloid ] -> [< `Landscape | `Portrait ] -> unit

beginprint_ext path mode format orientation opens and activates a print device with the given layout attributes.

The available formats are:

A4 0.210 x 0.297

B5 0.176 x 0.250

Letter 0.216 x 0.279

Legal 0.216 x 0.356

Executive 0.191 x 0.254

A0 0.841 x 1.189

A1 0.594 x 0.841

A2 0.420 x 0.594

A3 0.297 x 0.420

A5 0.148 x 0.210

A6 0.105 x 0.148

A7 0.074 x 0.105

A8 0.052 x 0.074

A9 0.037 x 0.052

B0 1.000 x 1.414

B1 0.500 x 0.707

B10 0.031 x 0.044

B2 0.500 x 0.707

B3 0.353 x 0.500

B4 0.250 x 0.353

B6 0.125 x 0.176

B7 0.088 x 0.125

B8 0.062 x 0.088

B9 0.044 x 0.062

C5E 0.163 x 0.229

Comm10E 0.105 x 0.241

DLE 0.110 x 0.220

Folio 0.210 x 0.330

Ledger 0.432 x 0.279

Tabloid 0.279 x 0.432

Parameters

pathname: Filename for the print device.

mode: Output mode (Color, GrayScale)

format: Output format (see table)

orientation: Page orientation (Landscape, Portait)