Module Gr.Workstation

type id

type workstation_type =

| WISS
Workstation Independent Segment ptr Storage

| WinGDI
Windows ptr GDI

| PS_1
PostScript (b/w @-> color)

| PS_2
PostScript (b/w @-> color)

| PS_3
PostScript (b/w @-> color)

| PS_4
PostScript (b/w @-> color)

| PDFPlain
Portable Document Format ptr plain

| PDFCompressed
Portable Document Format ptr compressed

| X_1
X ptr Windows

| X_2
X ptr Windows

| X_3
X ptr Windows

| X_4
X ptr Windows

| SunRF
Sun Raster file (RF)

| GIF87
Graphics Interchange Format ptr GIF87

| GIF89
Graphics Interchange Format ptr GIF89

| MotifUIL
Motif User Interface Language (UIL)

| BMP
Windows Bitmap (BMP)

| JPEG
JPEG image ptr file

| PNG
Portable Network Graphics file (PNG)

| TIFF
Tagged Image File Format (TIFF)

| Gtk
ptr Gtk

| Wx
ptr wxWidgets

| Qt4
ptr Qt4

| SVG
Scaleable Vector Graphics (SVG)

| WMF
Windows ptr Metafile

| Quartz
ptr Quartz

| Sock
Socket ptr driver

| ZMQ
0MQ ptr driver

| OGL
ptr OpenGL

Available workstation types, see also GR Workstation Types

val wid : int -> id

val open' : id -> string -> workstation_type -> unit

open' workstation_id connection type opens a graphical workstation.

Parameters

workstation_id: a workstation identifier

connection: a connection identifier

type: the desired workstation type

val close : id -> unit

val activate : id -> unit

val deactivate : id -> unit

val clear : unit -> unit

val update : unit -> unit

val set_window : float -> float -> float -> float -> unit

set_window xmin xmax ymin ymax sets the area of the NDC viewport that is to be drawn in the workstation window.

This function defines the rectangular area of the Normalized Device Coordinate space to be output to the device. By default, the workstation transformation will map the range [0,1] x [0,1] in NDC onto the largest square on the workstation’s display surface. The aspect ratio of the workstation window is maintained at 1 to 1.

Parameters

xmin: The left horizontal coordinate of the workstation window (0 <= xmin < xmax).

xmax: The right horizontal coordinate of the workstation window (xmin < xmax <= 1).

ymin: The bottom vertical coordinate of the workstation window (0 <= ymin < ymax).

ymax: The top vertical coordinate of the workstation window (ymin < ymax <= 1).

val set_viewport : float -> float -> float -> float -> unit

set_viewport xmin xmax ymin ymax defines the size of the workstation graphics window in meters.

This function places a workstation window on the display of the specified size in meters. This command allows the workstation window to be accurately sized for a display or hardcopy device, and is often useful for sizing graphs for desktop publishing applications.

Parameters

xmin: The left horizontal coordinate of the workstation window.

xmax: The right horizontal coordinate of the workstation window.

ymin: The bottom vertical coordinate of the workstation window.

ymax: The top vertical coordinate of the workstation window.

val copy_segment : segment -> unit

val redraw_segment : unit -> unit